requirements.txt
rich
protobuf
requests
GitPython
PyYAML
Jinja2
